Remove pkg/controller/chaosengine/types.go
